requirements.txt
# Used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask
pymongo[srv]
Flask-PyMongo
torch
python-dotenv
transformers>=4.12.5
py_vncorenlp==0.1.3
sentencepiece
tokenizer
underthesea
visen
